refactor: keep the notice log in its own file
The log is a different concern from deciding whether a reservation is
idle: it is read and written from the Slack side too, when the owner
answers that they are still going to use it. Giving it its own file lets
either side be read without the other.
