+++ src/application/usecases/mod.rs
@@ -76,7 +76,7 @@ pub mod update_resource_usage;
 pub use accept_reservation_proposal::AcceptReservationProposalUseCase;
 pub use create_resource_usage::CreateResourceUsageUseCase;
 pub use delete_resource_usage::DeleteResourceUsageUseCase;
-pub use detect_idle_reservations::{DetectIdleReservationsUseCase, IdleNoticeLog};
+pub use detect_idle_reservations::DetectIdleReservationsUseCase;
 pub use get_resource_usage_by_id::GetResourceUsageByIdUseCase;
 pub use grant_user_resource_access::GrantUserResourceAccessUseCase;
 pub use list_all_future_resource_usages::ListAllFutureResourceUsagesUseCase;
